Remove non-functional workflows and add a system test step
We inherited all of our github workflows from the repository template,
but the acceptance and deployment workflows currently do nothing. It
looks like the acceptance one is assuming an ephemeral environment for
the branch, which we would like to have at some point, but it's going to
be a while before it'll be ready.

I'm removing these now to avoid the confusion & waste of spinning up
runners that do nothing. We can recreate something similar when we're a
bit further along with the beta.

The test workflow now runs two kinds of tests: unit tests and system
tests.