Correctly link errors to the matching form field
We have a pattern for presenting errors at the top of the page. When you
click on the error, the relevant field should be focused (rather than
the previous behaviour of just navigating to the error)

In the case of checkboxes and radios, we need to link to the first
option. Since the ErrorList object we are iterating over doesn't know
whether the field contains checkboxes/radios, the easiest way to do this
is to always link to `ErrorList.field_id`, which is the ID generated by
`field.auto_id`, without a suffix like "-1", "-2" etc.

To make this all work, we have to explcitly set the `id` param of the
first checkbox/radio to the unsuffixed `field.auto_id`, while letting
the component generate the rest of the IDs with suffixes. If
https://github.com/nhsuk/nhsuk-frontend/pull/1112 is merged, then we
will no longer need to specify this, we can just provide the `idPrefix`