Add check in functionality to clinic show page
Allow CONFIRMED appointments to be checked in from this page. This
changes their status to CHECKED_IN.

This commit alters the method signature for the appointment_status macro
to allow a custom check_in_url for the form. It defaults to the
mammogram journey check in if no URL is provided.